addressing comments from first code review
