+++ hapi-fhir-jpaserver-model/src/main/java/ca/uhn/fhir/jpa/model/entity/PartitionablePartitionId.java
@@ -133,4 +133,8 @@ public class PartitionablePartitionId implements Cloneable {
 		}
 		return new PartitionablePartitionId(partitionId, theRequestPartitionId.getPartitionDate());
 	}
+
+	public static PartitionablePartitionId with(@Nullable Integer thePartitionId, @Nullable LocalDate thePartitionDate){
+		return new PartitionablePartitionId(thePartitionId, thePartitionDate);
+	}
 }
